:recycle: put web server config in `graphql` module
The `graphql::dpm` module was configuring global parameters
for the web server. This commit moves the global config up a
level into the `graphql` module. The `dpm` submodule configures
the portion of the config that pertains to the DPM API.

This organization will be useful when we add the parameter page API.